+++ src/soho/field-options/soho-field-options.directive.ts
@@ -1,0 +1,41 @@
+import {
+  AfterViewInit,
+  Directive,
+  ElementRef,
+  EventEmitter,
+  OnDestroy
+} from '@angular/core';
+
+@Directive({
+  selector: '[soho-field-options]' // tslint:disable-line
+})
+export class SohoFieldOptionsDirective implements AfterViewInit, OnDestroy {
+
+  private _options: SohoFieldOptionsSettings;
+
+  /**
+   * Local variables
+   */
+  private jQueryElement: JQuery;
+  private fieldOptions: SohoFieldOptionsStatic;
+
+  constructor(private element: ElementRef) {
+    this._options = {
+    };
+  }
+
+  ngAfterViewInit() {
+    this.jQueryElement = jQuery(this.element.nativeElement);
+    this.jQueryElement.fieldoptions(this._options);
+
+    this.fieldOptions = this.jQueryElement.data('field-options');
+  }
+
+  /** Destructor. */
+  ngOnDestroy() {
+    if (this.fieldOptions) {
+      this.fieldOptions.destroy();
+      this.fieldOptions = null;
+    }
+  }
+}