+++ src/soho/index.ts
@@ -15,6 +15,7 @@ export * from './dropdown/index';
 export * from './editor/index';
 export * from './error/index';
 export * from './expandablearea/index';
+export * from './field-options/index';
 export * from './fileupload/index';
 export * from './fileupload-advanced/index';
 export * from './hierarchy/index';